# optimize fft