| - |3|4|5|6|7|8|9|10|11|12|13|14|15|16|17|18|
| ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- |
| default | 1.8787 us 1.8841 us 1.8898 us | 27.871 us 29.831 us 32.287 us | 40.885 us 44.149 us 47.757 us | 49.677 us 52.739 us 56.196 us | 51.373 us 52.009 us 52.890 us | 100.27 us 106.84 us 114.88 us | 198.94 us 227.62 us 259.97 us | 277.42 us 283.34 us 290.65 us | 506.49 us 511.16 us 516.39 us | 1.0648 ms 1.0964 ms 1.1335 ms | 2.0286 ms 2.0736 ms 2.1223 ms | 5.6022 ms 5.8867 ms 6.2248 ms | 8.2442 ms 8.4402 ms 8.6559 ms | 19.105 ms 19.842 ms 20.658 ms | 37.900 ms 39.631 ms 41.765 ms | 73.686 ms 76.518 ms 80.451 ms |
| recursive | 510.74 ns 513.45 ns 516.32 ns | 13.491 us 13.600 us 13.712 us | 17.910 us 17.960 us 18.015 us | 29.320 us 29.443 us 29.582 us | 41.120 us 41.502 us 41.985 us | 54.564 us 54.695 us 54.837 us | 83.676 us 84.190 us 84.777 us | 144.14 us 145.33 us 146.65 us | 269.93 us 272.76 us 275.41 us | 534.74 us 539.85 us 544.27 us | 1.0254 ms 1.0290 ms 1.0332 ms | 2.0972 ms 2.1071 ms 2.1186 ms | 4.3206 ms 4.3327 ms 4.3451 ms | 9.1073 ms 9.1606 ms 9.2234 ms | 23.751 ms 24.573 ms 25.512 ms | 48.620 ms 50.373 ms 52.601 ms |