@NoCtrlZ 2022-04-21
default vs recursive
```
Gnuplot not found, using plotters backend
fft/k/3                 time:   [20.220 us 21.005 us 21.828 us]                     
                        change: [+946.96% +981.49% +1017.9%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
fft/k/4                 time:   [31.982 us 32.340 us 32.742 us]                     
                        change: [-5.5099% +1.3850% +8.2864%] (p = 0.70 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
fft/k/5                 time:   [53.212 us 54.160 us 55.314 us]                    
                        change: [+31.490% +42.701% +54.117%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
fft/k/6                 time:   [89.535 us 91.589 us 94.185 us]                    
                        change: [+56.984% +65.832% +74.838%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
fft/k/7                 time:   [140.74 us 144.76 us 149.66 us]                    
                        change: [+181.08% +193.50% +207.51%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
fft/k/8                 time:   [187.36 us 193.20 us 200.44 us]                    
                        change: [+75.673% +87.603% +99.821%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
fft/k/9                 time:   [299.71 us 312.57 us 328.41 us]                    
                        change: [+32.935% +44.048% +56.022%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
fft/k/10                time:   [367.88 us 370.73 us 374.20 us]                     
                        change: [+15.910% +20.701% +25.583%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
fft/k/11                time:   [579.59 us 584.90 us 590.63 us]                     
                        change: [+10.960% +15.524% +20.755%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
fft/k/12                time:   [1.0918 ms 1.1413 ms 1.1960 ms]                     
                        change: [-8.4797% -4.1487% +0.1571%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
Benchmarking fft/k/13: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.9s, enable flat sampling, or reduce sample count to 50.
fft/k/13                time:   [1.5596 ms 1.5726 ms 1.5869 ms]                      
                        change: [-23.747% -21.663% -19.410%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
fft/k/14                time:   [3.1042 ms 3.2624 ms 3.4586 ms]                      
                        change: [-48.550% -44.581% -39.810%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
fft/k/15                time:   [6.1783 ms 6.5865 ms 7.0714 ms]                     
                        change: [-27.118% -21.963% -15.926%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high severe
fft/k/16                time:   [12.283 ms 13.000 ms 13.875 ms]                     
                        change: [-38.823% -34.482% -29.886%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
fft/k/17                time:   [24.577 ms 25.168 ms 25.800 ms]                     
                        change: [-40.091% -36.495% -33.155%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  11 (11.00%) high mild
  5 (5.00%) high severe
Benchmarking fft/k/18: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.9s, or reduce sample count to 80.
fft/k/18                time:   [52.491 ms 53.694 ms 55.060 ms]                     
                        change: [-33.555% -29.828% -26.541%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  8 (8.00%) high mild
  7 (7.00%) high severe
```