@NoCtrlZ 2022-04-28
ifft -> arithmetic.rs
modularize butterfly arithmetic