@NoCtrlZ 2022-04-28
Can we avoid divisor for loop?