@NoCtrlZ 2022-04-28
`par_iter` > `scope`
https://docs.rs/rayon/latest/rayon/#how-to-use-rayon