@NoCtrlZ 2022-05-04
utilize parallelize in multicore