@NoCtrlZ 2022-05-04
refactor
https://github.com/zcash/halo2/blob/main/halo2_proofs/src/plonk/prover.rs#L296